# Conflicts:
#	app/src/main/java/com/example/calculatorprojectv2/LevelOneActivity.java

+++ app/src/test/java/com/example/calculatorprojectv2/QuestUnitTest.java
@@ -23,7 +23,7 @@ public class QuestUnitTest {
     public void quest_buttonLimit_isWorking() {
         for (int i = 0; i < 10000; i++) {
             Quest q = new Quest(i);
-            double[] testNums = q.getTestNums();
+            int[] testNums = q.getTestNums();
             // Checks to see if the length of each test number plus the length of the other test number plus one is the button limit
             // This is good for everything else
             assertEquals(String.valueOf(testNums[0]).length() + String.valueOf(testNums[1]).length() + 1,
@@ -36,7 +36,7 @@ public class QuestUnitTest {
         double precision = 0.01;
         for (int i = 0; i < 10000; i++) {
             Quest q = new Quest(i);
-            double[] testNums = q.getTestNums();
+            int[] testNums = q.getTestNums();
             switch (q.getOperationDesignation()) {
                 case 1:
                     assertEquals(testNums[0] + testNums[1], q.getTargetNumber(), precision);
@@ -66,7 +66,7 @@ public class QuestUnitTest {
     public void quest_numberLimit_isBelow() {
         for (int i = 0; i < 10000; i++) {
             Quest q = new Quest(i);
-            double[] testNums = q.getTestNums();
+            int[] testNums = q.getTestNums();
             assertTrue(testNums[0] < 10000);
             assertTrue(testNums[1] < 10000);
             assertTrue(q.getTargetNumber() < 10000);
